+++ polis-simulation/package.json
@@ -6,12 +6,13 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "ml-kmeans": "^6.0.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-range": "^1.10.0",
     "react-scripts": "5.0.1",
     "recharts": "^2.12.7",
-    "web-vitals": "^2.1.4",
-    "react-range": "^1.10.0"
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
