+++ polis-simulation/src/groupAwareConsensus.js
@@ -1,0 +1,98 @@
+// Required libraries
+const kmeans = require('node-kmeans'); // For K-means clustering
+const mathjs = require('mathjs');      // For matrix operations
+
+// Main function to process Polis data
+async function processPolisData(voteMatrix) {
+    // Step 1: K-means Clustering
+    const groups = await performClustering(voteMatrix);
+    
+    // Step 2: Calculate Group-Aware Consensus
+    const consensusScores = calculateGroupAwareConsensus(voteMatrix, groups);
+    
+    // Step 3: Rank Comments
+    const rankedComments = rankComments(consensusScores);
+    
+    return rankedComments;
+}
+
+// Step 1: K-means Clustering
+async function performClustering(voteMatrix) {
+    // Step 1a: Fine-grained clustering (K=100)
+    const fineGrainedClusters = await kMeansClustering(voteMatrix, 100);
+    
+    // Step 1b: Coarse-grained clustering (K=2 to 5)
+    let bestClusters = null;
+    let bestSilhouette = -Infinity;
+    
+    for (let k = 2; k <= 5; k++) {
+        const clusters = await kMeansClustering(voteMatrix, k);
+        const silhouette = calculateSilhouetteCoefficient(voteMatrix, clusters);
+        
+        if (silhouette > bestSilhouette) {
+            bestSilhouette = silhouette;
+            bestClusters = clusters;
+        }
+    }
+    
+    return bestClusters;
+}
+
+// Helper function for K-means clustering
+function kMeansClustering(data, k) {
+    return new Promise((resolve, reject) => {
+        kmeans.clusterize(data, {k: k}, (err, res) => {
+            if (err) reject(err);
+            else resolve(res);
+        });
+    });
+}
+
+// Calculate Silhouette Coefficient
+function calculateSilhouetteCoefficient(data, clusters) {
+    // Implementation of silhouette coefficient calculation
+    // This is a placeholder and should be implemented based on the specific requirements
+    return 0;
+}
+
+// Step 2: Calculate Group-Aware Consensus
+function calculateGroupAwareConsensus(voteMatrix, groups) {
+    const consensusScores = [];
+    
+    for (let commentIndex = 0; commentIndex < voteMatrix[0].length; commentIndex++) {
+        let consensusScore = 1;
+        
+        for (let group of groups) {
+            const groupVotes = group.clusterInd.map(index => voteMatrix[index][commentIndex]);
+            const agreeProbability = calculateAgreeProbability(groupVotes);
+            consensusScore *= agreeProbability;
+        }
+        
+        consensusScores.push({commentIndex, consensusScore});
+    }
+    
+    return consensusScores;
+}
+
+// Helper function to calculate agree probability
+function calculateAgreeProbability(votes) {
+    const agreeVotes = votes.filter(vote => vote === 1).length;
+    return (agreeVotes + 1) / (votes.length + 2); // Add 1 to numerator and 2 to denominator for Laplace smoothing
+}
+
+// Step 3: Rank Comments
+function rankComments(consensusScores) {
+    return consensusScores.sort((a, b) => b.consensusScore - a.consensusScore);
+}
+
+// Example usage
+const voteMatrix = [
+    [1, -1, 0, 1],
+    [-1, 1, 1, 0],
+    [1, 1, -1, 1],
+    [0, -1, 1, -1]
+];
+
+processPolisData(voteMatrix).then(rankedComments => {
+    console.log('Ranked Comments:', rankedComments);
+});
